Wait 15sec before disabling Services
When switching between two contexts using the service, it is briefly
paused before being started again. In some case (calendar) this implies
a full cache clear that can be costly to build back. Waiting 15 sec
should be more than enough time for the new context to bind to the
service again, preventing a pause.